src/requirements.txt
django==1.9.7
djangorestframework
markdown
django-filter
Pillow
pymysql